Rename gem to mandrillus
To avoid confusion with the original official version of this gem since
it appears unmaintained, I decide to rename the gem in order to publish it
on RubyGems. Mostly doing this so mandrill_mailer can switch to using this
gem instead of the original which has an unnecessarily strict requirement
for the json gem.